+++ src/components/Icon/BsIcon.ts
@@ -1,9 +1,10 @@
-import { useSizeHeight, useSizeStyles, useSizeWidth } from '@/components/Icon/mixins/iconApi';
+import { useSizeHeight, useSizeWidth } from '@/components/Icon/mixins/iconApi';
 import { iconProps, iconSizeObjectProp } from '@/components/Icon/mixins/iconProps';
 import type { TBsIcon, TBsIconSvg, TIconOptionProps } from '@/components/Icon/types';
 import { cssPrefix } from '@/mixins/CommonApi';
-import type { Prop } from 'vue';
-import { defineComponent, h } from 'vue';
+import type { TRecord } from '@/types';
+import Helper from '@/utils/Helper.ts';
+import { computed, defineComponent, h, type Prop } from 'vue';
 import BsIconSvg from './BsIconSvg';
 
 export default defineComponent<TBsIcon>({
@@ -14,21 +15,25 @@ export default defineComponent<TBsIcon>({
     },
     setup(props) {
         const thisProps = props as Readonly<TIconOptionProps>;
-        const szHeight = useSizeHeight(thisProps) as Prop<string>;
-        const szWidth = useSizeWidth(thisProps) as Prop<string>;
+        const szHeight = useSizeHeight(thisProps);
+        const szWidth = useSizeWidth(thisProps);
+        const styles = computed<TRecord>(() => ({
+            width: Helper.cssUnit(szWidth),
+            height: Helper.cssUnit(szHeight),
+        }));
 
         return () =>
             h(
                 'span',
                 {
                     class: [`${cssPrefix}icon`],
-                    style: useSizeStyles(thisProps),
+                    style: styles.value,
                 },
                 h<TBsIconSvg>(BsIconSvg, {
                     icon: props.icon,
                     filled: props.filled,
-                    height: szHeight,
-                    width: szWidth,
+                    height: szHeight as Prop<string>,
+                    width: szWidth as Prop<string>,
                     spin: props.spin,
                     pulse: props.pulse,
                     flip: props.flip,